Still need to get the flight data to display.

+++ db/create_airlines_table.sql
@@ -1,9 +1,9 @@
 create table Airlines(
     airline_id serial primary key,
-    air_freight integer,
-    fuel_surcharge integer,
-    security_surcharge integer,
-    screening integer,
+    air_freight numeric(6, 2),
+    fuel_surcharge numeric(6, 2),
+    security_surcharge numeric(6, 2),
+    screening numeric(6, 2),
     iata_airline_code integer,
     airline_name text,
     airline_type text
